requirements.txt
asgiref==3.3.4
Django==3.2.4
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
Pillow==8.2.0
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
